refactor: group webapp routes to allow opting out of header/footer layout
- also opt stream embed out of layout
