+++ src/routes/(app)/admin/events/+page.svelte
