+++ src/routes/(app)/event/[id]/+error.svelte
