+++ src/routes/(app)/event/[id]/+page.server.ts
