+++ src/routes/(app)/event/[id]/[slug]/+error.svelte
