+++ src/routes/(app)/event/[id]/[slug]/+page.ts
