+++ src/routes/(app)/event/[id]/[slug]/_embed_builder.svelte
