+++ src/routes/(app)/event/[id]/[slug]/embed/+page.ts
