+++ src/routes/(app)/event/[id]/[slug]/embed/_embed.svelte
