+++ src/routes/(app)/event/[id]/destroy/+page.server.ts
