+++ src/routes/(app)/event/[id]/edit/+page.svelte
